refactor(app): extract useForceGraphSimulation from MapPage
- Move spec loading, similarity weights/minSim state, the KNN graph
  derivation (node cache + cluster seeding), the settling gate
  (settled/tickProgress), search matching, and the tuned d3-force
  configuration into src/hooks/useForceGraphSimulation.ts
- Hook stays decoupled from ForceGraph2D's imperative ref API: it
  returns plain state + handlers (markSettled, handleEngineTick,
  resetWeights) and a static forceConfig; the only inversion is an
  onRepaint callback fired when thumbnails finish loading
- MapPage keeps the canvas element, paint callbacks, UI overlays, and
  all ref-coupled interaction handlers (camera fit, force wiring,
  fly-to); behavior, analytics events, aria attributes and keyboard
  interactions are unchanged
- outlierSquashForce + cluster color constants move to the hook module;
  MapPage.helpers stays a pure math module (header comment updated);
  hook exported from the src/hooks barrel
- New useForceGraphSimulation.test.ts (17 tests) covers graph
  derivation, weight/minSim handling, the settling gate, search
  matching, and the relocated outlierSquashForce suite; MapPage.test.tsx
  keeps the rendering/interaction tests (frontend suite 552 -> 562)

Part 10 of the frontend modernization roadmap.
